example/requirements.txt
rsanic==0.1.1
python-dotenv
asyncio_redis